+++ R/Methods.R
@@ -624,6 +624,9 @@ summary.fixest = function(object, vcov = NULL, cluster = NULL, ssc = NULL, .vcov
     # modifs of the table
     coeftable = cbind("Estimate" = object$coefficients, "Std. Error" = se_format,
                       "t value" = zvalue, "Pr(>|t|)" = pvalue)
+    if(object$method != "feols"){
+        colnames(coeftable) = colnames(object$coeftable)
+    }
 
     attr(coeftable, "type") = attr(se, "type") = attr(vcov, "type")